export file size as 4-byte value again

+++ Runtime/Scripts/GLTFSceneExporter.cs
@@ -835,7 +835,7 @@ namespace UnityGLTF
 			// write header
 			writer.Write(MagicGLTF);
 			writer.Write(Version);
-			writer.Write(glbLength);
+			writer.Write((int)glbLength);
 
 			gltfWriteJsonStreamMarker.Begin();
 			// write JSON chunk header.
